requirements.txt
dill==0.3.5.1
mlxpy==0.1
functorch==0.1.1
numpy==1.20.3
POT==0.8.2
torch==1.11.0+cu113
